microservices/Agent/requirements.txt
fastapi>=0.109.0
uvicorn==0.24.0
pydantic>=2.10,<3.0.0
python-dotenv==1.0.0
motor==3.3.1
pymongo==4.6.1
openai-agents==0.0.15
anyio>=4.5.0